requirements.txt
flask==2.1.1
pymongo==4.0.2
python-dotenv==1.0.0